requirements.txt
numpy~=1.21.5
h5py~=3.6.0
utils2p~=0.1.0
python-dotenv~=0.19.2
xarray~=0.20.2
matplotlib~=3.5.1
scipy~=1.7.3
pydantic~=1.9.0
dateparser~=1.1.0
tqdm~=4.62.3
pyyaml~=6.0
